# YJIT: Fix code GC freeing stubs with a trampoline