@maximecb 2022-12-15
This seems like a good change since it makes the stubs smaller.

But, can you explain why the stubs were not colocated with their code? Is it because the assembler can switch page in the middle of a block?