@maximecb 2022-12-15
> I like the code gen change. But the bug this is fixing and how this fixes it are just too subtle for me to review. Don't hold up the PR, but would love a walk through to understand this better.

IIUC it's because stubs don't always end up on the same page as the code jumping to them. So the stubs aren't always marked as live, meaning they can get collected.

If a stub gets collected while it's being called, then there's no more valid code to return to. This change makes it so that the trampoline part always remains live, so the address you return to is always valid.