@k0kubun 2022-12-15
> But, can you explain why the stubs were not colocated with their code? Is it because the assembler can switch page in the middle of a block?

It's kind of explained in https://github.com/ruby/ruby/pull/6933. On `invalidate_block_version`, we could generate stubs much later than the original inline code, not generating any inline code in new pages at that moment.