@XrXr 2022-12-15
> But, can you explain why the stubs were not colocated with their code? Is it because the assembler can switch page in the middle of a block?

The main way get this right now seems to be from invalidation regenerating branches with jumps to stubs. So for example a way this can happen is when we generate blocks A and B first, then we generate enough code to switch to a different page, and then we invalidate B. We generate a stub in the latest page and then patch A to jump to the stub.