@k0kubun 2022-12-15
Just in case, I confirmed that this doesn't have a negative impact on benchmarks too:

```
before: ruby 3.2.0dev (2022-12-14T18:21:46Z master 7a63114f8e) +YJIT [x86_64-linux]
last_commit=Remove dead code in get_next_shape_internal
after: ruby 3.2.0dev (2022-12-15T16:10:10Z yjit-stub-trampoline 30b3efca3d) +YJIT [x86_64-linux]

-------------  -----------  ----------  ----------  ----------  ------------  -------------
bench          before (ms)  stddev (%)  after (ms)  stddev (%)  before/after  after 1st itr
activerecord   64.7         0.5         64.5        0.6         1.00          1.00
hexapdf        1218.8       1.4         1203.2      1.0         1.01          1.00
liquid-render  65.8         1.3         66.2        1.2         0.99          1.01
mail           76.8         0.4         77.3        0.5         0.99          0.99
psych-load     979.0        0.1         981.6       0.1         1.00          1.00
railsbench     1074.9       1.7         1071.4      1.7         1.00          1.00
ruby-lsp       39.4         25.9        39.2        25.4        1.00          0.99
-------------  -----------  ----------  ----------  ----------  ------------  -------------
```